+++ server/budgeting/src/main/java/de/budgetfreak/budgeting/web/BudgetResource.java
@@ -1,5 +1,7 @@
 package de.budgetfreak.budgeting.web;
 
+import de.budgetfreak.utils.web.ResourceEmbeddableSupport;
+
 /**
  * Ressource representing a budget for a month.
  */
