+++ server/user-management/src/main/java/de/budgetfreak/usermanagement/web/UserControllerUserLinks.java
@@ -1,6 +1,7 @@
 package de.budgetfreak.usermanagement.web;
 
 import de.budgetfreak.usermanagement.domain.User;
+import de.budgetfreak.utils.web.ResourceLinks;
 import org.springframework.hateoas.Link;
 import org.springframework.stereotype.Component;
