+++ server/utils/web-tests/src/main/java/de/budgetfreak/utils/webtests/JsonHelper.java
@@ -1,4 +1,4 @@
-package de.budgetfreak.testutils.web;
+package de.budgetfreak.utils.webtests;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.core.JsonProcessingException;
