+++ server/utils/web/src/main/java/de/budgetfreak/utils/web/ResourceAssembler.java
@@ -1,4 +1,4 @@
-package de.budgetfreak.usermanagement.web;
+package de.budgetfreak.utils.web;
 
 import org.springframework.hateoas.Link;
 import org.springframework.hateoas.ResourceSupport;
