+++ settings.gradle
@@ -5,5 +5,5 @@ include 'server:user-management'
 include 'server:accounting'
 include 'server:budgeting'
 include 'server:application'
-include 'server:test-utils'
-include 'server:test-utils:web'
+include 'server:utils:web-tests'
+include 'server:utils:web'
